pyproject.toml
[project]
name = "decoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.30.0",
    "psycopg[binary]>=3.1.18",
    "bcrypt>=5.0.0",
    "pyjwt>=2.12.1",
    "redis>=7.4.0",
    "pydantic[email]>=2.13.1",
]
